+++ apps/epistola/src/test/kotlin/app/epistola/suite/EpistolaSuiteApplicationTests.kt
@@ -3,9 +3,11 @@ package app.epistola.suite
 import org.junit.jupiter.api.Test
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.context.annotation.Import
+import org.springframework.test.context.ActiveProfiles
 
 @Import(TestcontainersConfiguration::class)
 @SpringBootTest
+@ActiveProfiles("test")
 class EpistolaSuiteApplicationTests {
 
     @Test